third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-position-y-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-position-y value 'y-start'
  assert_true: 'y-start' is a supported value for background-position-y. expected true got false
[FAIL] Property background-position-y value 'y-end'
  assert_true: 'y-end' is a supported value for background-position-y. expected true got false
[FAIL] Property background-position-y value 'top -20%'
  assert_true: 'top -20%' is a supported value for background-position-y. expected true got false
[FAIL] Property background-position-y value 'bottom -10px'
  assert_true: 'bottom -10px' is a supported value for background-position-y. expected true got false
[FAIL] Property background-position-y value '0.5em, y-start, y-end'
  assert_true: '0.5em, y-start, y-end' is a supported value for background-position-y. expected true got false
[FAIL] Property background-position-y value 'calc(10px - 0.5em), top -20%, bottom 10px'
  assert_true: 'calc(10px - 0.5em), top -20%, bottom 10px' is a supported value for background-position-y. expected true got false
Harness: the test ran to completion.
